backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
google-auth
google-auth-oauthlib
google-auth-httplib2
pydantic[email]
email-validator
psycopg2-binary
python-multipart
watchfiles
sqlalchemy
aiofiles
pypdf
pytesseract
opencv-python
pdf2image
openai
google-generativeai
anthropic
python-docx
markdown
reportlab
fpdf2
websockets
